# api/openapi.yaml
openapi: 3.0.3
info:
  title: Feature Flag & Experimentation Platform API
  description: |
    Production-grade feature flag and experimentation platform API.

    This API provides endpoints for:
    - Organization, project, and environment management
    - Feature flag creation and management
    - User targeting and segmentation
    - A/B testing and experimentation
    - Analytics and results

    ## Authentication

    The API supports multiple authentication methods:
    - **Bearer Token**: JWT tokens for user sessions
    - **API Key**: API keys for programmatic access
    - **OIDC**: OpenID Connect for SSO integration

    ## Rate Limiting

    API endpoints are rate limited based on authentication:
    - User tokens: 1000 requests/minute
    - API keys: 10000 requests/minute
    - Anonymous: 100 requests/minute

  version: 1.0.0
  contact:
    name: Feature Flag Platform Team
    url: https://github.com/feature-flag-platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
  - url: https://api.feature-flags.dev/v1
    description: Development environment
  - url: https://api.feature-flags.com/v1
    description: Production environment
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      summary: User login
      description: Authenticate user with email and password
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      summary: Refresh token
      description: Refresh an expired JWT token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"

  # Organization endpoints
  /orgs:
    get:
      summary: List organizations
      description: Get all organizations the user has access to
      tags: [Organizations]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
                  pagination:
                    $ref: "#/components/schemas/Pagination"

    post:
      summary: Create organization
      description: Create a new organization
      tags: [Organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /orgs/{orgId}:
    parameters:
      - $ref: "#/components/parameters/OrgIdParam"

    get:
      summary: Get organization
      description: Get organization details
      tags: [Organizations]
      responses:
        "200":
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update organization
      description: Update organization details
      tags: [Organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete organization
      description: Delete an organization (owner only)
      tags: [Organizations]
      responses:
        "204":
          description: Organization deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"

  # Project endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Use "Bearer <api-key>" format
  parameters:
    OrgIdParam:
      name: orgId
      in: path
      required: true
      description: Organization ID
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        correlation_id:
          type: string
          description: Request correlation ID
    Pagination:
      type: object
      required: [page, limit, total, pages]
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        pages:
          type: integer
          description: Total number of pages
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Token expiry in seconds
        refresh_token:
          type: string
          description: Refresh token
        user:
          $ref: "#/components/schemas/User"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    # Core entity schemas
    Organization:
      type: object
      required: [id, name, slug, billing_tier, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
          maxLength: 100
        billing_tier:
          type: string
          enum: [free, pro, enterprise]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
...
    User:
      type: object
      required: [id, email, first_name, last_name]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        is_active:
          type: boolean
        email_verified:
          type: boolean
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    # Request schemas
    CreateOrganizationRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
          maxLength: 100
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
...